+++ lib/json_cache.dart
@@ -3,7 +3,6 @@ library json_cache;
 
 export 'src/json_cache.dart';
 export 'src/json_cache_cross_local_storage.dart';
-export 'src/json_cache_enc_prefs.dart';
 export 'src/json_cache_exception.dart';
 export 'src/json_cache_fake.dart';
 export 'src/json_cache_hive.dart';