+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
 - fix comment in JsonCacheHive class constructor — [120](https://github.com/dartoos-dev/json_cache/issues/120).
 
+### Removed
+
+- drop support for the 'EncryptedSharedPreferences' package — [125](https://github.com/dartoos-dev/json_cache/issues/125).
+
+
 ## [1.5.2] - 2023-02-28
+
 
 ### Changed